refactor & add_treap(dd)

+++ algorithm/bit_func/next_pow_2.cpp
@@ -1,17 +1,10 @@
-unsigned int nextPowerOf2(unsigned int n)
-{
-    unsigned count = 0;
-     
-    // First n in the below condition
-    // is for the case where n is 0
+long long  nextPowerOf2(long long n){
+    long long count = 0;
     if (n && !(n & (n - 1)))
         return n;
-     
-    while( n != 0)
-    {
+    while( n != 0){
         n >>= 1;
         count += 1;
     }
-     
     return 1 << count;
 }
